Add PW1's BGM wave archive to PW2 and have the copied music use it
Some of the copied music sounded strange in PW2, with very loud cymbals
(or something) instead of some soft tones.

Copying the whole wave archive of PW1 but not overwriting PW2's BGM
wave archive makes the ROM slightly bigger (~400KB) but assures that all
music sounds like it should.

Edit: this works in theory and makes the music sound well in Nitro
Studio 2 but in-game the music is silent. My guess is that the game
cannot load (because of memory) or simply doesn't want to load
WAVE_AGB_BGM_PW1. Keeping this commit for future reference.